requirements.txt
asgiref==3.4.1
Brotli==1.0.9
certifi==2021.10.8
charset-normalizer==2.0.10
Django==4.0
django-ckeditor==6.2.0
django-compression-middleware==0.4.1
django-cors-headers==3.11.0
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.4
django-environ==0.8.1
django-filter==21.1
django-js-asset==1.2.2
django-mptt==0.13.4
django-progressive-web-app==0.1.1
idna==3.3
Pillow==9.0.0
psycopg2==2.9.3
PyJWT==2.3.0
python-memcached==1.59
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
whitenoise==6.0.0
zstandard==0.17.0
